Move required field block definitions to utils
